Refactor SD card functions to return `int` for unified error handling and enhanced debugging.

+++ Project/hardware/inc/sdCard.h
@@ -11,15 +11,15 @@
 #endif
 #include <string.h>
 
-uint8_t sd_card_init();
+int sd_card_init();
 void sd_card_unmount();
 
-uint8_t sd_clear_score();
+int sd_clear_score();
 int sd_get_score();
-uint8_t sd_set_score(int score);
+int sd_set_score(int score);
 int sd_get_progress();
-uint8_t sd_set_progress(int progress);
+int sd_set_progress(int progress);
 
-uint8_t sd_get_locations(uint16_t type, char *buf, size_t *len, size_t max_len);
+int sd_get_locations(uint16_t type, char *buf, size_t *len, size_t max_len);
 
 #endif // SDCARD_H
